Split Raffle component into anothter component, UserInfo
